+++ Project/Assets/ML-Agents/Examples/3DBall/Scripts/Ball3DHardAgent.cs
@@ -13,7 +13,7 @@ public class Ball3DHardAgent : Agent
     public override void Initialize()
     {
         m_BallRb = ball.GetComponent<Rigidbody>();
-        m_ResetParams = Academy.Instance.GetSideChannel<FloatPropertiesChannel>();
+        m_ResetParams = SideChannelUtils.GetSideChannel<FloatPropertiesChannel>();
         SetResetParameters();
     }